Free up memory by deleting objects from memory
After 7 resets the game would freeze.
My hunch was that initializing new objects all the time was becoming too much.
This seemed to be the fix!

There's another bug that persists. After 7 resets, the player takes one
hit to go to game over.
I'm not sure why this is happening right now.
More than likely, it's a state issue.