Fix bug with 1 hit KOs + reset without deleting
The 1 KOs was caused by not recording the time with unsigned long.
With the value as int, it reached it's maximum value quickly.
That messed up the condition to allow multple hits at once.

The Player and Pedestrian objects are no longer pointers.
Instead of deleting them in memory, I just replace their values.
Less memory operations the better I believe